requirements.txt
numpy<2.0
opencv-python==4.7.0.72
Pillow==9.5.0
ollama==0.1.0
easyocr
networkx
matplotlib
requests
flask
